Transfert schedule logic to utils instead of in the UI
